Files managed, login window prototype created
