+++ c_cpp_properties.json
@@ -1,0 +1,42 @@
+{
+    "version": 4,
+    "configurations": [
+      {
+        "name": "Linux",
+        "includePath": [
+          "${workspaceFolder}/**",
+          "/usr/include/gtk-3.0",
+          "/usr/include/glib-2.0",
+          "/usr/lib/x86_64-linux-gnu/glib-2.0/include",
+          "/usr/include/pango-1.0",
+          "/usr/include/harfbuzz",
+          "/usr/include/gdk-pixbuf-2.0",
+          "/usr/include/cairo",
+          "/usr/include/pango-1.0",
+          "/usr/include/atk-1.0",
+          "/usr/include/gio-unix-2.0",
+          "/usr/include/gdk-3.0"
+        ],
+        "defines": [],
+        "compilerPath": "/usr/bin/gcc",
+        "cStandard": "c11",
+        "cppStandard": "c++17",
+        "intelliSenseMode": "linux-gcc-x64",
+        "browse": {
+          "path": [
+            "${workspaceFolder}/**",
+            "/usr/include/gtk-3.0",
+            "/usr/include/glib-2.0",
+            "/usr/lib/x86_64-linux-gnu/glib-2.0/include",
+            "/usr/include/pango-1.0",
+            "/usr/include/harfbuzz",
+            "/usr/include/gdk-pixbuf-2.0",
+            "/usr/include/cairo",
+            "/usr/include/atk-1.0",
+            "/usr/include/gio-unix-2.0",
+            "/usr/include/gdk-3.0"
+          ]
+        }
+      }
+    ]
+  }
