+++ client/inc/client.h
@@ -1,15 +1,9 @@
 #ifndef CLIENT_H
 #define CLIENT_H
 
-#include <gtk/gtk.h>
-#include <gtk/gtk.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <arpa/inet.h>
+/* This includes all necessary header files */
 
-#define MAIN_WINDOW_WIDTH 1280
-#define MAIN_WINDOW_HEIGHT 720
+#include "../inc/libs.h"
+#include "../inc/login.h"
 
 #endif
