+++ client/inc/libs.h
@@ -1,0 +1,14 @@
+#ifndef LIBS_H
+#define LIBS_H
+
+/* This header includes the necessary libraries and definitions */
+
+#include <gtk/gtk.h>
+#include <gtk/gtk.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <arpa/inet.h>
+
+#endif
