+++ client/inc/messenger.h
@@ -1,0 +1,11 @@
+#ifndef MESSENGER_H
+#define MESSENGER_H
+
+#include "../inc/libs.h"
+
+#define MAIN_WINDOW_WIDTH 1280
+#define MAIN_WINDOW_HEIGHT 720
+
+void create_main_window(GtkApplication* app, gpointer user_data);
+
+#endif
